LPS-58444 Move portal-impl's AggregateResourceBundle into portal-service
